apps/api/pyproject.toml
# apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smilegen-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "prisma>=0.13",
    "boto3>=1.34",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "pydantic-settings>=2.2",
    "httpx>=0.27",  # for test client
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
# conftest.py pins anyio_backend fixture to asyncio only (Prisma is asyncio-only)

[tool.hatch.envs.default.scripts]
dev = "uvicorn src.main:app --reload --port 8000"
test = "pytest tests/ -v"
